Moved some stuff into their own files (INCOMPLETE)
