+++ .gitignore
@@ -1,1 +1,2 @@
+src/__pycache__
 /.vscode
